+++ dots_server/db_models/uri_filtering_total_attack_connection_port.go
@@ -9,11 +9,11 @@ type UriFilteringTotalAttackConnectionPort struct {
 	PercentileType      string    `xorm:"'percentile_type' enum('LOW_PERCENTILE_L','MID_PERCENTILE_L','HIGH_PERCENTILE_L','PEAK_L') not null"`
 	Protocol            int       `xorm:"'protocol' not null"`
 	Port                int       `xorm:"'port' not null"`
-	Connection          int       `xorm:"connection"`
-	Embryonic           int       `xorm:"embryonic"`
-	ConnectionPs        int       `xorm:"connection_ps"`
-	RequestPs           int       `xorm:"request_ps"`
-	PartialRequestPs    int       `xorm:"partial_request_ps"`
+	Connection          uint64    `xorm:"connection"`
+	Embryonic           uint64    `xorm:"embryonic"`
+	ConnectionPs        uint64    `xorm:"connection_ps"`
+	RequestPs           uint64    `xorm:"request_ps"`
+	PartialRequestPs    uint64    `xorm:"partial_request_ps"`
 	Created             time.Time `xorm:"created"`
 	Updated             time.Time `xorm:"updated"`
 }
